Refactor DAO layer to use Spring's declarative transactions
This commit refactors the Data Access Object (DAO) layer of our application to use Spring's declarative transaction management. Each DAO method is now annotated with @Transactional, ensuring that the operations within each method are part of a single transaction. This eliminates the need for manual transaction management, simplifying the code and reducing boilerplate.

This commit also switches database connection from a local to a cloud server and updates the credentials to match, in preparation for deploying the application into a production environment. The former local properties have been kept but commented out for ease of use during local development.

In addition, the @Configuration annotation in CorsConfig file commented out as a temporary measure due to CORS issues encountered in a cloud environment.