Refactor code and added substring functionality in ChatVo
Removed unused and commented imports in WebConfig.java and UserController.java to improve code cleanliness. Also, specific imports were added instead of using a wildcard in UserController.java for transparency in dependencies used. Further, in ChatVo.java, a substring functionality was added to the 'setChatVo' method. This change was made to remove any characters that are unnecessary and could potentially impact the functionality or readability of the message.